+++ docs/conf.py
@@ -143,6 +143,6 @@ def linkcode_resolve(domain, info):
 
 
 extlinks = {
-    'issue': ('https://github.com/jcmgray/cotengra/issues/%s', 'GH %s'),
-    'pull': ('https://github.com/jcmgray/cotengra/pull/%s', 'PR %s'),
+    "issue": ("https://github.com/jcmgray/cotengra/issues/%s", "GH %s"),
+    "pull": ("https://github.com/jcmgray/cotengra/pull/%s", "PR %s"),
 }